+++ lua/config/options.lua
@@ -13,6 +13,7 @@ vim.opt.smartindent = true    -- Be smart about indentation
 vim.opt.wrap = true           -- Enable wrapping
 vim.opt.linebreak = true      -- Wrap lines at convenient points (words) NOT characters
 vim.opt.breakindent = true    -- Maintain indentation when wrapping
+vim.opt.hidden = true         -- Keep terminal jobs alive when their buffers/windows are hidden
 vim.opt.termguicolors = true  -- Enable true color support
 vim.opt.scrolloff = 8         -- Keep 8 lines of context around the cursor
 vim.opt.mouse = 'a'           -- Enable mouse support